Added options menu
The options menu is activated via the escape action when no other delegate is bound to the escape action (ie backing out of build mode). It pauses game time and can quit and be resumed from the menu. Has a stub button for options menu, but that is not implemented. Currently implementing save and load which also have buttons.